fix type mismatch between interface and declaration.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40084 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSTableView.m
@@ -170,7 +170,7 @@ typedef struct _tableViewFlags
 @end
 
 @interface NSTableView (EventLoopHelper)
-- (void) _trackCellAtColumn:(int)column row:(int)row withEvent:(NSEvent *)ev;
+- (void) _trackCellAtColumn:(NSInteger)column row:(NSInteger)row withEvent:(NSEvent *)ev;
 - (BOOL) _startDragOperationWithEvent:(NSEvent *)theEvent clickedRow:(NSUInteger)clickedRow;
 @end
